calendar 50%

+++ resources/views/calendar/full-calender.blade.php
@@ -61,39 +61,57 @@
             select:function(start, end, allDay)
             {
 
-                var title    = prompt('Event Title:');
-                var pasword  = prompt('Event password:');
-                var duration = prompt('Event Duration:');
 
 
 
-                if(title)
-                {
+                 // if(title) {
 
                     var start = $.fullCalendar.formatDate(start, 'Y-MM-DD HH:mm:ss');
 
                     var end = $.fullCalendar.formatDate(end, 'Y-MM-DD HH:mm:ss');
 
 
+                    // Check live DateTime
+                    var today = new Date();
+                    var date = today.getFullYear() + '-' + (today.getMonth() + 1) + '-' + today.getDate();
+                    var time = today.getHours() + ":" + today.getMinutes() + ":" + today.getSeconds();
+                    var LiveDateTime = date + ' ' + time;
 
-                    // alert(start)
-                    // alert(end)
-                    $.ajax({
-                        url:"{{ route('kk',app()->getLocale()) }}",
-                        type:"POST",
-                        data:{
-                            title: title,
-                            start: start,
-                            end: end,
-                            type: 'add'
-                        },
-                        success:function(data)
-                        {
-                            calendar.fullCalendar('refetchEvents');
-                            alert("Event Created Successfully");
+                    // compare
+
+                    var d1 = new Date(start);
+                    var d2 = new Date(LiveDateTime);
+
+                    if (d1 > d2) {
+
+                        var title    = prompt('Event Title:');
+                        var pasword  = prompt('Event password:');
+                        var duration = prompt('Event Duration:');
+
+                        if(title !== "" && pasword !== "" && duration !== ""){
+
+                        $.ajax({
+                            url: "{{ route('kk',app()->getLocale()) }}",
+                            type: "POST",
+                            data: {
+                                title: title,
+                                start: start,
+                                end: end,
+                                type: 'add'
+                            },
+                            success: function (data) {
+                                calendar.fullCalendar('refetchEvents');
+                                alert("Event Created Successfully");
+                            }
+                        });
+
+                        }else{
+                            alert('datarka');
                         }
-                    })
-                }
+                    }else{
+                        alert('You can not make appointments with հետ back date.');
+                    }
+
             },
             editable:true,
             eventResize: function(event, delta)
@@ -145,27 +163,42 @@
 
             eventClick:function(event)
             {
-                if(confirm("Are you sure you want to remove it?"))
-                {
-                    var id = event.id;
-                    $.ajax({
-                        url:"/en/full-calender/action",
-                        type:"POST",
-                        data:{
-                            id:id,
-                            type:"delete"
-                        },
-                        success:function(response)
-                        {
-                            calendar.fullCalendar('refetchEvents');
-                            alert("Event Deleted Successfully");
+                var id = event.id;
+                var user_id = event.user_id;
+
+
+                @if(!empty(Auth::user()->id))
+
+                    var AuthID = {{Auth::user()->id}}
+
+                    alert(user_id)
+                    if(AuthID === user_id) {
+
+
+                        if (confirm("Are you sure you want to remove it?")) {
+
+                            $.ajax({
+                                url: "/en/full-calender/action",
+                                type: "POST",
+                                data: {
+                                    id: id,
+                                    type: "delete"
+                                },
+                                success: function (response) {
+                                    calendar.fullCalendar('refetchEvents');
+                                    alert("Event Deleted Successfully");
+                                }
+                            })
                         }
-                    })
-                }
+                    }
+                @endif
+{{--                   @endif--}}
             }
         });
 
     });
+
+
 
 </script>
